# shared/openapi/family-cfo.v1.yaml
openapi: 3.1.0
info:
  title: Family CFO API
  version: 0.1.0
  description: >
    Initial specification for the Family CFO self-hosted backend API.
    This contract is the source of truth for generated Swift and Angular clients.
servers:
  - url: /api/v1
    description: Local Family CFO server
...
paths:
...
  /households:
    post:
      tags:
        - Household
      operationId: createHousehold
      summary: Bootstrap a household with its first owner (self-hosted first-run setup)
      responses:
        "201":
          description: Household created; returns an authenticated session for the new owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "409":
          $ref: "#/components/responses/Error"
...
  /income:
    get:
      tags:
        - Income
      operationId: listIncomeSources
      summary: List income sources for the household
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Income sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Income
      operationId: createIncomeSource
      summary: Create an income source
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeCreateRequest"
      responses:
        "201":
          description: Income source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeSource"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /income/{income_id}:
    patch:
      tags:
        - Income
      operationId: updateIncomeSource
      summary: Update an income source
      security:
        - bearerAuth: []
      parameters:
        - name: income_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeUpdateRequest"
      responses:
        "200":
          description: Income source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeSource"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags:
        - Income
      operationId: deleteIncomeSource
      summary: Delete an income source
      security:
        - bearerAuth: []
      parameters:
        - name: income_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Income source deleted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /chat/messages:
    post:
      tags:
        - Chat
      operationId: createChatMessage
      summary: Send a message to the financial advisor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "401":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Money:
      type: object
      required:
        - amount_minor
        - currency
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Signed integer amount in minor currency units.
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
...
    AuthSession:
      type: object
      required:
        - access_token
        - expires_at
        - household_id
        - user_id
        - role
      properties:
        access_token:
          type: string
        expires_at:
          type: string
          format: date-time
        household_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/HouseholdRole"
    HouseholdRole:
      type: string
      enum:
        - owner
        - adult
        - viewer
        - child
    RecurringFrequency:
      type: string
      enum:
        - weekly
        - biweekly
        - semimonthly
        - monthly
        - quarterly
        - annual
...
    IncomeSource:
      type: object
      required:
        - id
        - name
        - amount
        - frequency
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
...
    IncomeListResponse:
      type: object
      required:
        - income
      properties:
        income:
          type: array
          items:
            $ref: "#/components/schemas/IncomeSource"
...
    Recommendation:
      type: object
      required:
        - id
        - answer
        - assumptions
        - impacts
        - tradeoffs
        - alternatives
        - confidence
        - calculation_refs
      properties:
        id:
          type: string
          format: uuid
        answer:
          type: string
        assumptions:
          type: array
          items:
            type: string
        impacts:
          type: array
          items:
            $ref: "#/components/schemas/Impact"
        tradeoffs:
          type: array
          items:
            type: string
        alternatives:
          type: array
          items:
            type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        calculation_refs:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    Impact:
      type: object
      required:
        - area
        - summary
      properties:
        area:
          type: string
          enum:
            - cash_flow
            - emergency_fund
            - debt
            - savings_goal
            - retirement
            - net_worth
            - other
        summary:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
    ChatRequest:
      type: object
      required:
        - message
      properties:
        conversation_id:
          type: string
          format: uuid
        message:
          type: string
          minLength: 1
          maxLength: 4000
    ChatResponse:
      type: object
      required:
        - conversation_id
        - recommendation
      properties:
        conversation_id:
          type: string
          format: uuid
        recommendation:
          $ref: "#/components/schemas/Recommendation"
...
    IncomeCreateRequest:
      type: object
      required:
        - name
        - amount
        - frequency
      properties:
        name:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
    IncomeUpdateRequest:
      type: object
      properties:
        name:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
...